Revert "PR for HARP 2.0"
