--- examples/midi_pitch_shifter/.gitignore
@@ -1,1 +1,0 @@
-_outputs
