--- examples/midi_pitch_shifter/README.md
@@ -1,13 +1,0 @@
----
-title: Midi Pitch Shifter
-emoji: 🏆
-colorFrom: indigo
-colorTo: gray
-sdk: gradio
-sdk_version: 4.42.0
-app_file: app.py
-pinned: false
-license: mit
----
-
-Check out the configuration reference at https://huggingface.co/docs/hub/spaces-config-reference
