--- examples/midi_pitch_shifter/requirements.txt
@@ -1,3 +1,0 @@
--e git+https://github.com/TEAMuP-dev/pyharp.git#egg=pyharp
-torch
-torchaudio
