+++ examples/templates/wav2wav/app.py
@@ -25,7 +25,6 @@ def process_fn(input_audio_path):
 
     Returns:
         output_audio_path (str): the filepath of the processed audio.
-        output_labels (LabelList): any labels to display.
     """
 
     """
@@ -36,7 +35,7 @@ def process_fn(input_audio_path):
 
     """
     <YOUR AUDIO PROCESSING CODE HERE>
-    # Perform a trivial operation (i.e. gain)
+    # Perform a trivial operation (i.e. boosting)
     signal.audio_data = 2 * signal.audio_data
     """
 
@@ -46,13 +45,7 @@ def process_fn(input_audio_path):
     output_audio_path = save_audio(signal, None)
     """
 
-    """
-    <YOUR LABELING CODE HERE>
-    # No output labels
-    output_labels = LabelList()
-    """
-
-    return output_audio_path, output_labels
+    return output_audio_path
 
 
 # Build Gradio endpoint
@@ -62,6 +55,7 @@ with gr.Blocks() as demo:
         # <YOUR UI ELEMENTS HERE>
     ]
 
+    # Build endpoint
     app = build_endpoint(model_card=model_card,
                          components=components,
                          process_fn=process_fn)
