+++ pyharp/__init__.py
@@ -1,3 +1,1 @@
 from .core import *
-from .labels import *
-from .media import *
