--- pyharp/media/__init__.py
@@ -1,2 +1,0 @@
-from .audio import *
-from .midi import *
